third_party/WebKit/LayoutTests/external/wpt/cookies/secure/create-cookie-http-expected.txt

This is a testharness.js-based test.
FAIL Secure cookies cannot be set by insecure origins assert_equals: Insecure cookies are always present expected "0.2512442783278579" but got ""
Harness: the test ran to completion.
